server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4
    # via openai
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
blinker==1.6.2
    # via flask
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   flask
    #   gtts
colorama==0.4.6
    # via
    #   click
    #   tqdm
flask==2.3.1
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.11.0
    # via google-cloud-texttospeech
google-auth==2.17.3
    # via google-api-core
google-cloud-texttospeech==2.14.1
    # via -r requirements.in
googleapis-common-protos==1.59.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.54.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.54.0
    # via google-api-core
gtts==2.3.1
    # via -r requirements.in
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.6.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
openai==0.27.4
    # via -r requirements.in
proto-plus==1.22.2
    # via google-cloud-texttospeech
protobuf==4.22.3
    # via
    #   google-api-core
    #   google-cloud-texttospeech
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydub==0.25.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.28.2
    # via
    #   google-api-core
    #   gtts
    #   openai
    #   speechrecognition
rsa==4.9
    # via google-auth
simpleaudio==1.0.4
    # via -r requirements.in
six==1.16.0
    # via google-auth
speechrecognition==3.10.0
    # via -r requirements.in
tqdm==4.65.0
    # via openai
urllib3==1.26.15
    # via requests
werkzeug==2.3.1
    # via flask
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via importlib-metadata
